+++ inst/app/conditionalUI.R
@@ -1,66 +1,84 @@
 # Functions to render the UI based on other input
 
-output$ui.idvar <- renderUI({
+output$ui_id <- renderUI({
 
   req(data())
 
   if (is.null(data()$file)) return(NULL)
 
   choices <- colnames(data()$file)
-  selectInput('dea.idvar', 'ID variable', choices = choices, multiple = FALSE)
+  selectInput('dea_id', 'Firm ID', choices = choices, multiple = FALSE)
 
 })
 
-output$ui.inputs <- renderUI({
+output$ui_inputs <- renderUI({
 
   req(data())
 
   if (is.null(data()$file)) return(NULL)
 
   # Restore input if we are restoring previous state
-  if (!is.null(restoreVals$inputs))
-    selected <- restoreVals$inputs
-  else
-    selected <- NULL
-
+  selected <- if (!is.null(restoreVals$inputs)) restoreVals$inputs else NULL
   choices <- data()$cols[sapply(data()$file, is.numeric, USE.NAMES = FALSE)]
-  selectInput('dea.input', 'Inputs', choices = choices, selected = selected, multiple = TRUE)
+
+  selectInput('dea_input', 'Inputs', choices = choices, selected = selected, multiple = TRUE)
 
 })
 
-output$ui.outputs <- renderUI({
+observeEvent(input$dea_input, {
+  selected_inputs <- input$dea_input
+  selected <- input$dea_output
+  choices <- data()$cols[sapply(data()$file, is.numeric, USE.NAMES = FALSE)]
+  if (length(selected_inputs) > 0) {
+    choices <- choices[!(choices %in% selected_inputs)]
+  }
+  if (!is.null(selected) && any(selected %in% selected_inputs)) {
+    selected <- selected[!(selected %in% selected_inputs)]
+  }
+  updateSelectInput(session, 'dea_output', choices = choices, selected = selected)
+})
+
+output$ui_outputs <- renderUI({
 
   req(data())
 
   if (is.null(data()$file)) return(NULL)
 
   # Restore input if we are restoring previous state
-  if (!is.null(restoreVals$outputs))
-    selected <- restoreVals$outputs
-  else
-    selected <- NULL
-
+  selected <- if (!is.null(restoreVals$outputs)) restoreVals$outputs else NULL
   choices <- data()$cols[sapply(data()$file, is.numeric, USE.NAMES = FALSE)]
-  if (!is.null(input$dea.input) && !any(input$dea.input == ''))
-    choices <- choices[!(choices %in% c(input$datafile, input$dea.input))]
-  selectInput('dea.output', 'Outputs', choices = choices, selected = selected, multiple = TRUE)
+
+  selectInput('dea_output', 'Outputs', choices = choices, selected = selected, multiple = TRUE)
 
 })
 
-output$ui.timeseries <- renderUI({
+observeEvent(input$dea_output, {
+  selected_outputs <- input$dea_output
+  selected <- input$dea_input
+  choices <- data()$cols[sapply(data()$file, is.numeric, USE.NAMES = FALSE)]
+  if (length(selected_outputs) > 0) {
+    choices <- choices[!(choices %in% selected_outputs)]
+  }
+  if (!is.null(selected) && any(selected %in% selected_outputs)) {
+    selected <- selected[!(selected %in% selected_outputs)]
+  }
+  updateSelectInput(session, 'dea_input', choices = choices, selected = selected)
+})
 
-  req(data()$file, input$dea.idvar)
+output$ui_timeseries <- renderUI({
 
-  if (!(input$dea.idvar %in% colnames(data()$file))) return(NULL)
+  req(data()$file, params()$id)
+
+  if (!(params()$id %in% colnames(data()$file))) return(NULL)
 
   # If the ID variable is not unique, we assume time series
-  value <- any(duplicated(data()$file[, input$dea.idvar]))
+  value <- any(duplicated(data()$file[, params()$id]))
 
   checkboxInput('hasyear', 'Time series data', value = value)
 
 })
 
-output$ui.year <- renderUI({
+output$ui_year <- renderUI({
 
   req(data(), input$hasyear)
 
@@ -86,11 +104,11 @@ output$ui.year <- renderUI({
 
   choices <- data()$cols[which(year_vars)] #names(chk_min > 1900 & chk_max < 2100)
   selected <- ifelse(length(choices) == 1, choices[[1]], NULL)
-  selectInput('dea.year', 'Year variable', choices = choices, selected = selected, multiple = TRUE)
+  selectInput('dea_year', 'Year variable', choices = choices, selected = selected, multiple = TRUE)
 
 })
 
-output$ui.subset <- renderUI({
+output$ui_subset <- renderUI({
 
   req(data()$file)
 
@@ -100,7 +118,7 @@ output$ui.subset <- renderUI({
 
 })
 
-output$ui.subset.info <- renderUI({
+output$ui_subset_info <- renderUI({
 
   req(data()$file)